backend/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
